+++ release_notes.md
@@ -1,3 +1,7 @@
+### Fixed
+
+- Cannot anonymize instance lacking `PatientSex` ([#24](https://github.com/blairconrad/dicognito/issues/24))
+
 ### Additional Items
 
 - Test on a CI environment ([#11](https://github.com/blairconrad/dicognito/issues/11))